+++ CHANGELOG.md
@@ -6,10 +6,21 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- towncrier release notes start -->
-## 0.24.7 (2025-06-15)
+## 0.24.9 (2025-06-15)
 
 ### Improvements
 - Added support for mapping as part of ocean integration mappings.
+
+## 0.24.8 (2025-06-11)
+
+### Bug Fixes
+- Fixed missing syncing state metrics reporting during resource processing
+
+## 0.24.7 (2025-06-11)
+
+### Bug Fixes
+- Update is oauth enabled condition to check if a path to oauth token is set.
+- Update requests.
 
 ## 0.24.6 (2025-06-09)